+++ examples/account/evmapp/src/main/java/io/horizen/examples/EvmAppModule.java
@@ -61,6 +61,7 @@ public class EvmAppModule extends AccountAppModule {
 
         // It's integer parameter that defines slot duration. The minimum valid value is 10, the maximum is 300.
         int consensusSecondsInSlot = 12;
+        String appVersion = "";
 
         // use a custom object which implements the stopAll() method
         SidechainAppStopper applicationStopper = new EvmAppStopper();
@@ -104,5 +105,9 @@ public class EvmAppModule extends AccountAppModule {
         bind(Integer.class)
                 .annotatedWith(Names.named("ConsensusSecondsInSlot"))
                 .toInstance(consensusSecondsInSlot);
+
+        bind(String.class)
+                .annotatedWith(Names.named("AppVersion"))
+                .toInstance(appVersion);
     }
 }