+++ qa/SidechainTestFramework/secure_enclave_http_api_server.py
@@ -5,6 +5,8 @@ import subprocess
 
 from flask import Flask, request, json
 
+from test_framework.util import assert_equal
+
 
 class SecureEnclaveApiServer(object):
 
@@ -26,6 +28,9 @@ class SecureEnclaveApiServer(object):
         @self.app.route('/api/v1/createSignature', methods=['POST'])
         def sign_message():
             content = json.loads(request.data)
+            if "akka-http" in request.headers['User-Agent']:
+                assert_equal("application/json", request.headers['Content-Type'])
+                assert_equal("application/json", request.headers['Accept'])
             logging.info("SecureEnclaveApiServer /api/v1/createSignature received request " + str(content))
             if 'privateKey' not in content:
                 pk = content['publicKey']
@@ -83,7 +88,7 @@ def launch_signing_tool(json_parameters):
 
     java_ps = subprocess.Popen(["java", "-jar",
                                 os.getenv("SIDECHAIN_SDK", "..")
-                                + "/tools/signingtool/target/sidechains-sdk-signingtools-0.7.0.jar",
+                                + "/tools/signingtool/target/sidechains-sdk-signingtools-0.7.1.jar",
                                 "createSignature", json_param], stdout=subprocess.PIPE)
     db_tool_output = java_ps.communicate()[0]
     try: