+++ sdk/src/main/scala/io/horizen/account/AccountAppModule.scala
@@ -51,7 +51,8 @@ abstract class AccountAppModule extends com.google.inject.AbstractModule {
            @Named("CustomMessageProcessors") customMessageProcessors: JList[MessageProcessor],
            @Named("ApplicationStopper") applicationStopper : SidechainAppStopper,
            @Named("ForkConfiguration") forkConfigurator : ForkConfigurator,
-           @Named("ConsensusSecondsInSlot") secondsInSlot: Int
+           @Named("ConsensusSecondsInSlot") secondsInSlot: Int,
+           @Named("AppVersion") appVersion: String
          ): AccountSidechainApp = {
     synchronized {
       if (app == null) {
@@ -65,7 +66,8 @@ abstract class AccountAppModule extends com.google.inject.AbstractModule {
           applicationStopper,
           forkConfigurator,
           chainInfo,
-          secondsInSlot
+          secondsInSlot,
+          appVersion
         )
       }
     }