+++ sdk/src/main/scala/io/horizen/params/RegTestParams.scala
@@ -40,7 +40,8 @@ case class RegTestParams(
                           override val sidechainCreationVersion: SidechainCreationVersion = SidechainCreationVersion1,
                           override val chainId: Long = 1111111,
                           override val isCSWEnabled: Boolean = true,
-                          override val isNonCeasing: Boolean = false
+                          override val isNonCeasing: Boolean = false,
+                          override val isHandlingTransactionsEnabled: Boolean = true
                         ) extends NetworkParams {
   override val EquihashN: Int = 48
   override val EquihashK: Int = 5